split MetricBuilder
