+++ core/src/main/java/com/jetbrains/tmp/learning/stepik/StepikConnectorGet.java
@@ -28,6 +28,10 @@ import com.jetbrains.tmp.learning.courseFormat.Lesson;
 import com.jetbrains.tmp.learning.courseFormat.Section;
 import com.jetbrains.tmp.learning.courseFormat.Task;
 import com.jetbrains.tmp.learning.courseFormat.TaskFile;
+import com.jetbrains.tmp.learning.stepik.metric.MetricActions;
+import com.jetbrains.tmp.learning.stepik.metric.MetricBuilder;
+import com.jetbrains.tmp.learning.stepik.metric.MetricsWrapper;
+import com.jetbrains.tmp.learning.stepik.metric.PluginNames;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpStatus;
 import org.apache.http.NameValuePair;
@@ -189,9 +193,9 @@ public class StepikConnectorGet {
     }
 
     public static Course getCourse(@NotNull final Project project, @NotNull final CourseInfo info) {
-        MetricBuilder.MetricsWrapper metric = MetricBuilder.getInstance()
-                .addTag(MetricBuilder.PluginNames.STEPIK_UNION)
-                .addTag(MetricBuilder.MetricActions.GET_COURSE)
+        MetricsWrapper metric = MetricBuilder.getInstance()
+                .addTag(PluginNames.STEPIK_UNION)
+                .addTag(MetricActions.GET_COURSE)
                 .setCourseId(info.id)
                 .build();
         StepikConnectorPost.postMetric(metric);
